refactor: split out different functions in printer
