+++ src/Backend/MoonrakerMessageTypes.swift
@@ -13,5 +13,4 @@ enum MoonrakerMethod: String {
     case printerEmergency_stop = "printer.emergency_stop"
     case printerRestart = "printer.restart"
     case printerFirmware_restart = "printer.firmware_restart"
-    //case notifyProcStatUpdate
 }
